Implemented the GitHub API calls for all well-defined data-points from the excel sheet
It is now possible to request:
- Basic repository data
- Basic version data
- Basic owner data
- Contributor count
- GitStar ranking
- Yearly commit count

Do keep in mind that the contributor count and GitStar ranking are subject to caching, so they might be slightly inaccurate. Both also return higher number than the GitHub website itself, but I can't find the reason for it.

There is also no error-checking yet. At the moment everything presumes that the previous calls have been successful.