Created a controller and JSON schema
The controller will continuously try to read the console for input.
This input should be of the form that is described in 'input_json_structure.json'. For clarity, and example has also been added.

The controller will request the given data-points from the corresponding website (and thus API endpoint/spider).

When the controller has all the data, it will return a JSON object containing the wanted data-points. If it was not able to find a data-point, it will return None as the value of the data-point.
The output has the form described in 'output_json_structure.json'. For clarity, an example has also been included here.

The JSON structure files contain all possible data-points.