Bug fix
Since the last commit, the scan would always return None. This was due to a wrong path check, and due to a mistake in updating the output JSON.

The program should now run as intended.

Keep in mind however, that we only check to see if the UNIX socket exists. This does not necessarily mean that the scanner is also running. A better solution for this still has to be found.