Update test_controller.py
Last commit I made a mistake in setting up the testing framework for the Controller class. Now, I have changed the functions and descriptions to be what they were supposed to be.

For the TestControllerRun functions, you only need to mock the function calls to the other Controller functions (get_github_data needs to be mocked for test_run_gh).

For the TestControllerData functions, you will need to mock all of the functions that can be called (aka the functions that actually grab the data)